+++ src/main/java/com/pact/rentcar/model/dto/request/AddVehicleRequest.java
@@ -10,9 +10,7 @@ public class AddVehicleRequest {
     private String registration;
     private String brand;
     private String model;
-    private BigDecimal dailyFee;
-
-
+    private Double dailyFee;
 
 
 }
